backend/requirements.txt
fastapi
uvicorn
websockets
google-generativeai
beautifulsoup4
requests
python-dotenv
feedparser
